requirements.txt
Flask
flask-cors
flask-restful
gunicorn
requests
huggingface_hub
python-dotenv
openai